refactor: login page in components
